Fix missing GHC.Generics support for GHC 7.4 (re #13)
GHC.Generics support was added via #9 but left out support for
GHC 7.4, thereby resulting in an *observable* conditional API, which in turn
violates the PVP contract and cause problems down the line for packages
depending on `data-default-class` which depend on `Generics` support.

This simple patch adds `Generics`-support for GHC 7.4 and consequently
fixes #13.